+++ packages/typography/src/index.ts
@@ -22,10 +22,10 @@ export { default as Label } from './Label/Label';
 export { labelTypeScaleStyles } from './Label/Label.styles';
 export type { LabelProps } from './Label/Label.types';
 export { default as Link } from './Link/Link';
-export { anchorClassName } from './Link/Link.styles';
 export type { LinkProps } from './Link/Link.types';
 export { default as Overline } from './Overline/Overline';
 export type { OverlineProps } from './Overline/Overline.types';
+export { anchorClassName } from './SharedLink/SharedLink.styles';
 export { bodyTypeScaleStyles } from './styles';
 export { default as Subtitle } from './Subtitle/Subtitle';
 export type { SubtitleProps } from './Subtitle/Subtitle.types';
